+++ features/step_definitions/custom_web_steps.rb
@@ -74,7 +74,7 @@ end
 
 And /^I want to mention (?:him|her) from the profile$/ do
   find('#mention_button').click
-  within('#facebox') do
+  within('#mentionModal') do
     click_publisher
   end
 end
@@ -126,6 +126,12 @@ end
 
 When /^(.*) in the modal window$/ do |action|
   within('#facebox') do
+    step action
+  end
+end
+
+When /^(.*) in the mention modal$/ do |action|
+  within('#mentionModal') do
     step action
   end
 end