+++ features/support/paths.rb
@@ -28,8 +28,8 @@ module NavigationHelpers
       when /^"([^\"]*)"'s page$/
         p = User.find_by_email($1).person
         { path: person_path(p),
-          # '.diaspora_handle' on desktop, '.description' on mobile
-          special_elem: { selector: '.diaspora_handle, .description', text: p.diaspora_handle }
+          # '#diaspora_handle' on desktop, '.description' on mobile
+          special_elem: { selector: '#diaspora_handle, .description', text: p.diaspora_handle }
         }
       when /^"([^\"]*)"'s photos page$/
         p = User.find_by_email($1).person